[QUAR-51] ♻️ Refactor: 페이지별 커스텀훅 제거하여 각페이지에서 상태관리로 구조 개선, 제네릭 타입 지정 제거, 유추할 수 있는 상태명 주석 제거
